backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.22
requests==2.31.0
pydantic==2.4.2
langchain-core==0.1.28
langchain-community==0.0.21
python-dotenv==1.0.0
langchain-google-genai==1.0.10
beautifulsoup4==4.12.2
